Rearranges clients package to use better names
