+++ packages/clients/client.go
@@ -1,4 +1,4 @@
-package models
+package clients
 
 type Client struct {
 	Id        string
