+++ packages/clients/clirepo/client_db_repository.go
@@ -1,8 +1,8 @@
-package repositories
+package clirepo
 
 import (
 	"github.com/juju/errors"
-	"theodo.red/creditcompanion/packages/clients/models"
+	"theodo.red/creditcompanion/packages/clients"
 	"theodo.red/creditcompanion/packages/database"
 	"theodo.red/creditcompanion/packages/database/marsh"
 	"theodo.red/creditcompanion/packages/database/tdynamo"
@@ -12,8 +12,8 @@ type ClientDBRepository struct {
 	repo database.Repository
 }
 
-func (r *ClientDBRepository) Get(id string) (*models.Client, error) {
-	client := new(models.Client)
+func (r *ClientDBRepository) Get(id string) (*clients.Client, error) {
+	client := new(clients.Client)
 	err := r.repo.GetByUniqueField("id", id, client)
 	if err != nil {
 		return nil, errors.Annotate(err, "Call to repo failed.")
@@ -21,8 +21,8 @@ func (r *ClientDBRepository) Get(id string) (*models.Client, error) {
 	return client, nil
 }
 
-func (r *ClientDBRepository) GetByEmail(email string) (*models.Client, error) {
-	client := new(models.Client)
+func (r *ClientDBRepository) GetByEmail(email string) (*clients.Client, error) {
+	client := new(clients.Client)
 	err := r.repo.GetByUniqueField("email", email, client)
 	if err != nil {
 		return nil, errors.Annotate(err, "Call to repo failed.")
@@ -30,7 +30,7 @@ func (r *ClientDBRepository) GetByEmail(email string) (*models.Client, error) {
 	return client, nil
 }
 
-func NewDynamoClientRepository(db tdynamo.DynamoDbInterface) models.ClientRepository {
+func NewDynamoClientRepository(db tdynamo.DynamoDbInterface) ClientRepository {
 	repo := new(ClientDBRepository)
 	marshallingRepo := marsh.NewMarshallingDynamoRepository(db, "clients")
 	repo.repo = marshallingRepo
