+++ packages/clients/clirepo/client_repository.go
@@ -1,0 +1,8 @@
+package clirepo
+
+import "theodo.red/creditcompanion/packages/clients"
+
+type ClientRepository interface {
+	Get(id string) (*clients.Client, error)
+	GetByEmail(email string) (*clients.Client, error)
+}
