--- packages/clients/models/client_repository.go
@@ -1,6 +1,0 @@
-package models
-
-type ClientRepository interface {
-	Get(id string) (*Client, error)
-	GetByEmail(email string) (*Client, error)
-}
